Progress with UN metadata imports.

+++ scripts/batch/import_global_metadata.py
@@ -4,6 +4,9 @@ import re
 import unicodedata
 import urllib
 import os.path
+import PyPDF2
+import glob
+import yaml
 from zipfile import ZipFile
 
 """
@@ -19,26 +22,36 @@ Other metadata: The other global metadata is only available in English, so this
 script only grabs the UN metadata for English.
 """
 
+def find_between(s, start, end):
+  if start not in s or end not in s:
+    return None
+  between = (s.split(start))[1].split(end)[0]
+  return between.strip()
+
 def sdg_number_from_text(text):
   """
   This parses a string of text and pulls out the SDG number. Possible formats of
   return value are: '1', '1.1', '1.1.1'
   """
   if pd.isnull(text):
-    return 'was null'
+    return None
 
   matches = re.findall(r'(\d+)(\.\w+)?(\.\w+)?', text)
   if len(matches) > 0:
     return ''.join(matches[0])
   else:
-    return 'no matches'
+    return None
 
 def sdg_text_without_number(text, number):
   """
   This simply removes a number from some text.
   """
   normalized = unicodedata.normalize("NFKD", str(text))
-  return normalized.replace(number, '').strip()
+  parts = normalized.split(number)
+  if len(parts) == 2:
+    return parts[1].lstrip('.').strip()
+  else:
+    return normalized
 
 def main():
   global_goals = {}
@@ -67,6 +80,7 @@ def main():
       'skipfooter': 6,
       'encoding': 'utf-8',
     }
+
     df = pd.read_excel(spreadsheet_url, **import_options)
     for index, row in df.iterrows():
       # If the 'indicator' column in empty, this is a Goal.
@@ -90,12 +104,79 @@ def main():
 
   # Next get the (Engish only, currently) metadata.
   zip_filename = 'SDG-indicator-metadata.zip'
-  if not os.path.isfile(zip_filename):
-    remote_url = 'https://unstats.un.org/sdgs/metadata/files/SDG-indicator-metadata.zip'
-    urllib.request.urlretrieve(remote_url, zip_filename)
-  with ZipFile(zip_filename, 'r') as zip:
-    # printing all the contents of the zip file
-    zip.printdir()
+  example_pdf = 'Metadata-01-01-01a.pdf'
+  unzip_folder = 'temp-import-files'
+  example_pdf_path = os.path.join(unzip_folder, example_pdf)
+  if not os.path.isfile(example_pdf_path):
+    if not os.path.isfile(zip_filename):
+      remote_url = 'https://unstats.un.org/sdgs/metadata/files/SDG-indicator-metadata.zip'
+      urllib.request.urlretrieve(remote_url, zip_filename)
+      with ZipFile(zip_filename, 'r') as zip:
+        zip.extractall(unzip_folder)
+
+  for filepath in glob.iglob(unzip_folder + '/*.pdf'):
+    # Figure out the indicator number.
+    indicator = filepath.replace(unzip_folder, '')
+    indicator = indicator.replace(os.sep + 'Metadata-', '')
+    indicator = indicator.replace('.pdf', '')
+    indicator_parts = indicator.split('-')
+    indicator = '.'.join(map(lambda x: x.lstrip('0').lower(), indicator_parts))
+
+    if indicator not in global_indicators['en']:
+      global_indicators['en'][indicator] = {}
+
+    pdfFileObject = open(filepath, 'rb')
+    pdfReader = PyPDF2.PdfFileReader(pdfFileObject)
+    count = pdfReader.getNumPages()
+    pdfContent = ''
+    for i in range(count):
+      page = pdfReader.getPage(i)
+      pdfContent += page.extractText()
+
+    # Next we use a series of very fragile tricks for getting certain metadata.
+    # Mostly we look for text on either side of what we're looking for.
+    organization = find_between(pdfContent, 'Organization(s):', 'Concepts and definitions')
+    if organization:
+      organizations = organization.split('\n \n')
+      organizations = list(map(lambda x: x.replace('\n', ''), organizations))
+      organization = '\n'.join(organizations)
+      organization = organization.strip()
+    if organization and organization != '':
+      global_indicators['en'][indicator]['custodian_agency'] = organization
+
+    description = find_between(pdfContent, 'Definition:', 'Rationale:')
+    if description:
+      description = description.replace('\n', '').strip()
+    if description and description != '':
+      global_indicators['en'][indicator]['description'] = description
+
+    # Just guess at the remote link from the filepath.
+    remote_folder = 'https://unstats.un.org/sdgs/metadata/files/'
+    remote_link = filepath.replace(unzip_folder + os.sep, remote_folder)
+    global_indicators['en'][indicator]['metadata_link'] = remote_link
+
+  # Finally merge the results into the YAML files.
+  all_results = {
+    'global_goals.yml': global_goals,
+    'global_targets.yml': global_targets,
+    'global_indicators.yml': global_indicators,
+  }
+  for yaml_filename in all_results:
+    for language in all_results[yaml_filename]:
+      translation_path = os.path.join('translations', language, yaml_filename)
+      yaml_data = None
+      with open(translation_path, 'r') as stream:
+        yaml_data = yaml.load(stream)
+      if not yaml_data:
+        yaml_data = {}
+      for item in all_results[yaml_filename][language]:
+        if item not in yaml_data:
+          yaml_data[item] = all_results[yaml_filename][language][item]
+        else:
+          for key in all_results[yaml_filename][language][item]:
+            yaml_data[item][key] = all_results[yaml_filename][language][item][key]
+      with open(translation_path, 'w') as outfile:
+        yaml.dump(yaml_data, outfile, default_flow_style=False, allow_unicode=True)
 
 if __name__ == '__main__':
   main()
